+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 Changes:
 
   * Fix MinION sequencing container models (was previously reporting the sequencing kit instead)
+  * Added sequencing kit field for Oxford Nanopore runs
 
 # 1.9.0